Add Grav 1.4 to images
It occurred to me that we will need more than just “latest” sooner,
rather than later, due to the fact that we’re building dependencies on
this and having our dependencies be version locked will be important.
Right now we’re not “version locking” an image to patch version only
major and minor. There is probably a better way to do this, such as
templates, etc. but I don’t have time to look into them right now.